get books from api - make code more testable
